rtc-scraper/requirements.txt
selenium
requests
pillow
webdriver-manager
psycopg2-binary